renamed main folder, split taxo into thematic files and fix few imports
