+++ src/wsl_library/pdfextraction/main.py
@@ -3,47 +3,51 @@ import os
 
 import ollama
 
-from wslibrary.domain import taxonomy
-from wslibrary.pdfextraction.llm import prompts
-from wslibrary.pdfextraction.llm.ollama_extraction import extract_ollama_from_paper
-from wslibrary.pdfextraction.pdf.pymu import get_pymupdf4llm
+from wsl_library.domain import (
+    paper_taxonomy,
+    publication_taxonomy,
+    geographical_taxonomy,
+    themes_taxonomy,
+)
+from wsl_library.pdfextraction.llm import prompts
+from wsl_library.pdfextraction.llm.ollama_extraction import extract_ollama_from_paper
+from wsl_library.pdfextraction.pdf.pymu import get_pymupdf4llm
 
 
-OLLAMA_MODELS  = [m["model"] for m in ollama.list().get("models", [])]
+OLLAMA_MODELS = [m["model"] for m in ollama.list().get("models", [])]
 PROMPT_FCTS = {name: obj for name, obj in prompts.__dict__.items() if callable(obj)}
-TAXS = {name: obj for name, obj in taxonomy.__dict__.items() if callable(obj)}
+TAXS = {name: obj for name, obj in paper_taxonomy.__dict__.items() if callable(obj)}
+
 
 def get_args_parser():
     parser = argparse.ArgumentParser(description="Extract taxonomy from a paper pdf file")
     parser.add_argument("--pdf-path", type=str, help="Path to the pdf file")
     parser.add_argument(
-        "--model",
-        type=str,
-        choices=OLLAMA_MODELS,
-        help="Model to use for extraction"
+        "--model", type=str, choices=OLLAMA_MODELS, help="Model to use for extraction"
     )
     parser.add_argument(
         "--taxonomy",
         type=str,
         choices=list(TAXS.keys()),
         default="Paper",
-        help="Taxonomy to use for extraction"
+        help="Taxonomy to use for extraction",
     )
     parser.add_argument(
         "--prompt-type",
         type=str,
         choices=list(PROMPT_FCTS.keys()),
         default="main_parts_prompt",
-        help="Type of prompt to use for extraction"
+        help="Type of prompt to use for extraction",
     )
     parser.add_argument(
         "--output-folder",
         type=str,
         default=None,
-        help="Name of the folder where the extracted information will be stored"
+        help="Name of the folder where the extracted information will be stored",
     )
 
     return parser.parse_args()
+
 
 def get_files(path: str) -> tuple[str, list[str]]:
     """
@@ -56,7 +60,7 @@ def get_files(path: str) -> tuple[str, list[str]]:
     if path.endswith(".pdf"):
         pdf_name = path.split("/")[-1]
         return path.replace(pdf_name, ""), [pdf_name]
-    
+
     elif os.path.isdir(path):
         pdfs = [
             os.path.join(dir_name.replace(path, ""), file)
@@ -67,11 +71,12 @@ def get_files(path: str) -> tuple[str, list[str]]:
 
         if len(pdfs) == 0:
             raise ValueError("There are no pdfs in the indicated folder.")
-        
+
         return path, pdfs
-    
+
     else:
         raise ValueError("The indicated path is not a pdf or a folder.")
+
 
 def get_folders(pdf_files: list[str]) -> set[str]:
     """
@@ -80,17 +85,17 @@ def get_folders(pdf_files: list[str]) -> set[str]:
         pdf_files (list[str]): List of pdf files.
     Returns:
         set[str]: Set of the folders that will be created
-    
+
     """
     return set(["/".join(file.split("/")[:-1]) for file in pdf_files])
 
-     
+
 def main():
     """
     Main function to extract the taxonomy from a pdf files.
     The function will find the pdfs in the indicated path, extract the text from them,
     make a prompt from the text and extract the taxonomy from the prompt.
-    
+
     For each pdf file 3 files are created in the output folder:
     - The extracted text
     - The prompt used for the extraction
@@ -111,7 +116,7 @@ def main():
 
     else:
         raise ValueError(f"{args.prompt_type} has not been implemented")
-    
+
     # get the pdfs at the indicated path
     input_path, pdfs = get_files(input_path)
 
@@ -133,17 +138,13 @@ def main():
         results[pdf]["prompt"] = prompt
 
         # extract the taxonomy from the extracted texts
-        paper = extract_ollama_from_paper(
-            prompt,
-            model_name,
-            tax
-        )
+        paper = extract_ollama_from_paper(prompt, model_name, tax)
         results[pdf]["tax"] = paper
-    
+
     # make the needed dirs and save the information
     for f in folders:
         os.makedirs(output_path + f, exist_ok=True)
-    
+
     for pdf, c in results.items():
         # save the extracted text
         text_path = output_path + pdf.replace(".pdf", ".md")
