+++ tests/pdfextraction/pdf/test_pymu.py
@@ -4,7 +4,7 @@
 import os
 from pathlib import Path
 
-from wslibrary.pdfextraction.pdf.pymu import get_pymupdf4llm
+from wsl_library.pdfextraction.pdf.pymu import get_pymupdf4llm
 
 
 HERE_PATH = Path(__file__).parent
